Separate geographic selection stuff into a separate file
